requirements.txt
email-validator
flask
flask-bcrypt
flask-sqlalchemy
flask-login
flask-migrate
flask-wtf
python-dotenv
wtforms
wtforms-sqlalchemy
werkzeug
